requirements.txt
python-dotenv
schedule
paho-mqtt
selenium
bs4
pingparsing